Make providers configurable and fix testing
