--- mex/common/db/models.py
@@ -1,3 +1,0 @@
-from sqlalchemy.orm import declarative_base
-
-Base = declarative_base()
