+++ mex/common/identity/types.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class IdentityProvider(Enum):
+    """Choice of available identity providers."""
+
+    BACKEND = "backend"
+    DUMMY = "dummy"
